refactor: extract media persistence from AgentWorkspace
